Cannot save a bookmark that is exactly the same

+++ src/js/map/bookmarks.js
@@ -19,15 +19,37 @@ function readData () {
 
 function saveBookmark (newBookmark) {
     let currentData = readData();
-    currentData.data.push(newBookmark);
-    LocalStorage.setItem(STORAGE_BOOKMARKS_KEY, JSON.stringify(currentData));
-    return true;
+
+    // If the bookmarkExists, then return false because we did not add it to the list
+    if (bookmarkExists(newBookmark)) {
+        return false;
+    }
+    // Otherwise, add the bookmark to LocalStorage and return true
+    else {
+        currentData.data.push(newBookmark);
+        LocalStorage.setItem(STORAGE_BOOKMARKS_KEY, JSON.stringify(currentData));
+        return true;
+    }
 }
 
 function clearData () {
     LocalStorage.setItem(STORAGE_BOOKMARKS_KEY, JSON.stringify(DEFAULT_BOOKMARKS_OBJECT));
     EventEmitter.dispatch('clearbookmarks', {});
     return true;
+}
+
+function bookmarkExists (newBookmark) {
+    let currentData = readData();
+    for (var bookmark of currentData.data) {
+        // If the bookmark exists in the saved list, then return true
+        if (bookmark.label === newBookmark.label && bookmark.lat === newBookmark.lat &&
+            bookmark.lng === newBookmark.lng && bookmark.zoom === newBookmark.zoom) {
+            return true;
+        }
+    }
+
+    // If bookmark does not exist in current list return false
+    return false;
 }
 
 let bookmarks = {
